Configuration to .Rprofile
Before this change, there was an issue with the Dockerfile when trying to install R libraries using commands like system("conda install r-terra"). While the libraries were technically installed, they were not placed in the correct location.  I confirmed this by checking with .libPaths().

After discovering the problem, I searched for a solution and found that by adding a .Rprofile file with the correct paths, the installation now works as expected. This modification sets the CONDA_PREFIX and updates the PATH in the .Rprofile to ensure R libraries are installed in the appropriate environment